+++ src/interfaces/ILiquidationPool.sol
@@ -2,5 +2,5 @@
 pragma solidity ^0.8.17;
 
 interface ILiquidationPool {
-    function distributeFees(uint256 _amount) external;
+    function distributeFees(uint _amount) external;
 }
