+++ src/interfaces/ILiquidationPoolManager.sol
@@ -1,15 +1,15 @@
 // SPDX-License-Identifier: UNLICENSED
 pragma solidity ^0.8.17;
 
-import "src/interfaces/ITokenManager.sol";
+import 'src/interfaces/ITokenManager.sol';
 
 interface ILiquidationPoolManager {
     struct Asset {
         ITokenManager.Token token;
-        uint256 amount;
+        uint amount;
     }
 
     function distributeFees() external;
 
-    function runLiquidation(uint256 _tokenId) external;
+    function runLiquidation(uint _tokenId) external;
 }
