+++ src/interfaces/INFTMetadataGenerator.sol
@@ -1,10 +1,13 @@
 // SPDX-License-Identifier: UNLICENSED
 pragma solidity 0.8.17;
 
-import "src/interfaces/ISmartVault.sol";
+import 'src/interfaces/ISmartVault.sol';
 
 interface INFTMetadataGenerator {
-    function generateNFTMetadata(uint256 _tokenId, ISmartVault.Status memory _vaultStatus)
+    function generateNFTMetadata(
+        uint _tokenId,
+        ISmartVault.Status memory _vaultStatus
+    )
         external
         view
         returns (string memory);
