+++ src/interfaces/ISmartVaultIndex.sol
@@ -2,8 +2,8 @@
 pragma solidity 0.8.17;
 
 interface ISmartVaultIndex {
-    function getTokenIds(address _user) external view returns (uint256[] memory);
-    function getVaultAddress(uint256 _tokenId) external view returns (address payable);
-    function addVaultAddress(uint256 _tokenId, address payable _vault) external;
-    function transferTokenId(address _from, address _to, uint256 _tokenId) external;
+    function getTokenIds(address _user) external view returns (uint[] memory);
+    function getVaultAddress(uint _tokenId) external view returns (address payable);
+    function addVaultAddress(uint _tokenId, address payable _vault) external;
+    function transferTokenId(address _from, address _to, uint _tokenId) external;
 }
